Moved creation of commands present after init into caller responsibility
